assignment/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-default-group==1.2.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.4
djangorestframework==3.12.2
idna==2.10
include==0.2.2
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==8.0.1
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tabulate==0.8.7
task==0.2.0
uritemplate==3.0.1
urllib3==1.26.2
